[4.6.3] Added support for CraftEngine 26.4+ version and drop support for version below that.         Added {number:<number>} built-in placeholder.         The support for CoinEngine plugin will be removed in future update, please replace it to ExcellentEconomy in your server as soon as possible.

+++ core/src/main/java/cn/superiormc/mythictotem/objects/AbstractSingleRun.java
@@ -106,11 +106,11 @@ public abstract class AbstractSingleRun {
     }
 
     public double getDouble(String path) {
-        return MathUtil.doCalculate(section.getString(path));
+        return MathUtil.doCalculate(section.getString(path)).doubleValue();
     }
 
     public double getDouble(String path, Player player, AbstractThingData data) {
-        return MathUtil.doCalculate(replacePlaceholder(section.getString(path), player, data));
+        return MathUtil.doCalculate(replacePlaceholder(section.getString(path), player, data)).doubleValue();
     }
 
     public boolean getBoolean(String path, boolean defaultValue) {
